+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Releases
 0.3.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Allow binary values to be passed for ``string`` types.
 
 
 0.3.0 (2015-09-09)